Fix cython binding
Add SpaceMapUtil lib in cython/opencog/CMakeList.txt and
update the initNewMap().

+++ opencog/cython/opencog/spacetime.pxd
@@ -22,7 +22,7 @@ cdef extern from "opencog/spacetime/SpaceServer.h" namespace "opencog":
         cSpaceServer(cAtomSpace&)
         const cSpaceMap& getMap(cHandle)
         cHandle addOrGetSpaceMap(octime_t, string, double, string)
-        bool addSpaceInfo(cHandle, cHandle, bool, bool, octime_t, int, int, int, string)
+        bool addSpaceInfo(cHandle, cHandle, bool, bool, octime_t, double, double, double, string)
         void removeSpaceInfo(cHandle, cHandle, octime_t, string)
         void clear()
         void setTimeServer(cTimeServer*)
